Crate embassy_executor_macros

embassy-executor-macros

An Embassy project.

NOTE: Do not use this crate directly. The macros are re-exported by embassy-executor.

Attribute Macros

main_cortex_m

Creates a new executor instance and declares an application entry point for Cortex-M spawning the corresponding function body as an async task.

main_riscv

Creates a new executor instance and declares an application entry point for RISC-V spawning the corresponding function body as an async task.

main_std

Creates a new executor instance and declares an application entry point for STD spawning the corresponding function body as an async task.

main_wasm

Creates a new executor instance and declares an application entry point for WASM spawning the corresponding function body as an async task.

task

Declares an async task that can be run by embassy-executor. The optional pool_size parameter can be used to specify how many concurrent tasks can be spawned (default is 1) for the function.